+++ docs/Changelog.md
@@ -7,6 +7,8 @@ See https://github.com/cowwoc/requirements.java/commits/master for a full list.
 * Improvements
     * When referencing to a variable number of elements, always use the plural form of a type (e.g.
       `actual must contain X entries` even if `X` is one).
+    * Added special handling for `Collection.contains(Object)` throwing `NullPointerException` because the
+      collection does not permit null elements.
 
 ## Version 10.0 - 2024/10/31